fix:set Long for Id in Person, fix it in PatientService+DoctorService

+++ src/main/java/com/emelyan/models/Person.java
@@ -1,8 +1,6 @@
 package com.emelyan.models;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
@@ -16,7 +14,7 @@ import java.time.LocalDate;
 public  class Person {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Integer id;
+    private Long id;
 
     private String name;
     private String surname;
